+++ packages/cli/src/lib/check-lockfile.test.ts
