+++ packages/cli/src/lib/check-lockfile.ts
