+++ packages/cli/src/lib/check-version.test.ts
