+++ packages/cli/src/lib/check-version.ts
