+++ packages/cli/src/lib/config.ts
