+++ packages/cli/src/lib/flags.test.ts
