+++ packages/cli/src/lib/log.ts
