+++ packages/cli/src/lib/shell.ts
