+++ packages/cli/src/lib/virtual-routes.test.ts
