+++ packages/cli/src/lib/virtual-routes.ts
